+++ elasticsearch_dsl/__init__.py
@@ -84,7 +84,7 @@ from .update_by_query import UpdateByQuery
 from .utils import AttrDict, AttrList, DslBase
 from .wrappers import Range
 
-VERSION = (8, 0, 0)
+VERSION = (8, 12, 0)
 __version__ = VERSION
 __versionstr__ = ".".join(map(str, VERSION))
 __all__ = [